fix common/browser component layering
